requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.1.0
Django==6.0
django-allauth==65.13.1
django-formtools==2.5.1
django-otp==1.7.0
django-phonenumber-field==8.4.0
django-rest-authtoken==2.1.4
django-two-factor-auth==1.18.1
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
gunicorn==25.0.0
idna==3.11
jwt==1.4.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
slippers==0.6.2
sqlparse==0.5.5
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
